bulk fix module imports error after refactor
