+++ src/components/configure_run.py
@@ -18,7 +18,10 @@ from typing import NamedTuple
 from kfp.v2 import dsl
 from kfp.v2.dsl import Artifact
 from kfp.v2.dsl import Output
-from src import config
+
+import sys
+sys.path.append('..')
+import config
 
 
 @dsl.component(
