+++ src/components/hhblits.py
@@ -19,7 +19,9 @@ from kfp.v2.dsl import Artifact
 from kfp.v2.dsl import Input
 from kfp.v2.dsl import Output
 
-from src import config
+import sys
+sys.path.append('..')
+import config
 
 
 @dsl.component(
