+++ src/components/hmmsearch.py
@@ -17,7 +17,10 @@ from kfp.v2 import dsl
 from kfp.v2.dsl import Artifact
 from kfp.v2.dsl import Input
 from kfp.v2.dsl import Output
-from src import config
+
+import sys
+sys.path.append('..')
+import config
 
 
 @dsl.component(
