+++ src/components/relax_protein.py
@@ -18,7 +18,9 @@ from kfp.v2.dsl import Artifact
 from kfp.v2.dsl import Input
 from kfp.v2.dsl import Output
 
-from src import config
+import sys
+sys.path.append('..')
+import config
 from typing import List
